sa-logic/sa/requirements.txt
Werkzeug==0.16.1
Flask==0.12.2
textblob==0.15.0
kafka-python==2.0.2
